requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic[email]==2.12.5
email-validator>=2.0.0
requests==2.31.0
fhirclient==4.3.0
python-multipart
Pillow


